Some more cleanups
